+++ python/fnllm/fnllm/errors.py
@@ -6,8 +6,8 @@ fnllm Custom Errors.
 All custom errors for fnllm are re-exported from here.
 """
 
+from .base.services.errors import FailedToGenerateValidJsonError, RetriesExhaustedError
 from .caching.blob import InvalidBlobCacheArgumentsError, InvalidBlobContainerNameError
-from .services.errors import FailedToGenerateValidJsonError, RetriesExhaustedError
 from .tools.errors import ToolInvalidArgumentsError, ToolNotFoundError
 
 __all__ = [