+++ python/fnllm/fnllm/openai/llm/openai_text_chat_llm.py
@@ -10,7 +10,9 @@ from typing import TYPE_CHECKING, Any, cast
 from openai.types.chat.chat_completion_message_param import ChatCompletionMessageParam
 from typing_extensions import Unpack
 
-from fnllm.base.base import BaseLLM
+from fnllm.base.base_llm import BaseLLM
+from fnllm.openai.services.history_extractor import OpenAIHistoryExtractor
+from fnllm.openai.services.usage_extractor import OpenAIUsageExtractor
 from fnllm.openai.types.aliases import OpenAIChatCompletionModel, OpenAIChatModel
 from fnllm.openai.types.chat.io import (
     OpenAIChatCompletionInput,
@@ -18,20 +20,17 @@ from fnllm.openai.types.chat.io import (
     OpenAIChatOutput,
 )
 from fnllm.openai.types.chat.parameters import OpenAIChatParameters
+from fnllm.openai.utils import build_chat_messages
 from fnllm.types.metrics import LLMUsageMetrics
 
-from .services.history_extractor import OpenAIHistoryExtractor
-from .services.usage_extractor import OpenAIUsageExtractor
-from .utils import build_chat_messages
-
 if TYPE_CHECKING:
+    from fnllm.base.services.cache_interactor import Cached, CacheInteractor
+    from fnllm.base.services.json import JsonHandler
+    from fnllm.base.services.rate_limiter import RateLimiter
+    from fnllm.base.services.retryer import Retryer
+    from fnllm.base.services.variable_injector import VariableInjector
     from fnllm.events.base import LLMEvents
     from fnllm.openai.types.client import OpenAIClient
-    from fnllm.services.cache_interactor import Cached, CacheInteractor
-    from fnllm.services.json import JsonHandler
-    from fnllm.services.rate_limiter import RateLimiter
-    from fnllm.services.retryer import Retryer
-    from fnllm.services.variable_injector import VariableInjector
     from fnllm.types.generics import TJsonModel
     from fnllm.types.io import LLMInput