+++ python/fnllm/fnllm/openai/llm/services/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-
-"""Package with OpenAI specific features to be used to wrap base LLM protocol interfaces."""