+++ python/fnllm/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+* Move `fnllm.services` to `fnllm.base.services`
+* Move `fnllm.config` to `fnllm.base.config`
+* Move `fnllm.openai.llm.services` to `fnllm.openai.services`
+* Add `fnllm.enums` to re-export common enumerations.
 ### Deprecated
 ### Removed
 ### Fixed